+++ sickrage/core/searchers/proper_searcher.py
@@ -277,6 +277,7 @@ class srProperSearcher(object):
                 result.version = curProper.version
                 result.content = curProper.content
                 result.size = curProper.provider._get_size(curProper.url)
+                result.files = curProper.provider._get_files(curProper.url)
 
                 # snatch it
                 snatchEpisode(result, SNATCHED_PROPER)